requirements.txt
pyyaml==6
click==8